utils/image_preprocess.py
from PIL import Image

def crop_image_with_box(img, box):
    """
    Crops the image based on the given bounding box coordinates.
    
    Parameters:
    img (PIL.Image): A PIL Image object.
    box (list): Bounding box coordinates [x_min, y_min, x_max, y_max] as proportions.
    
    Returns:
    cropped_image: Cropped image as a PIL Image object.
    """
    width, height = img.size

    # Convert proportional coordinates to pixel coordinates
    left = int(box[0] * width)
    top = int(box[1] * height)
    right = int(box[2] * width)
    bottom = int(box[3] * height)

    # Crop the image using the calculated coordinates
    cropped_image = img.crop((left, top, right, bottom))

    return cropped_image
    