requirements.txt
pillow
datasets
torch
torchvision
